// frontend/web/src/services/API/proserve-wb-projects-api/models/GetProjectEnrolmentsResponseItem.ts
export interface GetProjectEnrolmentsResponseItem {
    /**
     * 
     * @type {string}
     * @memberof GetProjectEnrolmentsResponseItem
     */
    id?: string;
    /**
     * User TID
     * @type {string}
     * @memberof GetProjectEnrolmentsResponseItem
     */
    userId?: string;
    /**
     * User e-mail
     * @type {string}
     * @memberof GetProjectEnrolmentsResponseItem
     */
    userEmail?: string;
    /**
     * Project ID
     * @type {string}
     * @memberof GetProjectEnrolmentsResponseItem
     */
    projectId?: string;
    /**
     * Enrolment status
     * @type {string}
     * @memberof GetProjectEnrolmentsResponseItem
     */
    status?: string;
    /**
     * Enrolment ticket ID
     * @type {string}
     * @memberof GetProjectEnrolmentsResponseItem
     */
    ticketId?: string;
    /**
     * Enrolment approver
     * @type {string}
     * @memberof GetProjectEnrolmentsResponseItem
     */
    approver?: string;
    /**
     * Enrolment decline reason
     * @type {string}
     * @memberof GetProjectEnrolmentsResponseItem
     */
    reason?: string;
    /**
     * Point in time where enrolment is created
     * @type {string}
     * @memberof GetProjectEnrolmentsResponseItem
     */
    createDate?: string;
    /**
     * Point in time where enrolment is resolved
     * @type {string}
     * @memberof GetProjectEnrolmentsResponseItem
     */
    resolveDate?: string;
    /**
     * System in which enrolment request was approved
     * @type {string}
     * @memberof GetProjectEnrolmentsResponseItem
     */
    sourceSystem?: string;
}